anvil_zksync_core::formatter

Trait PrettyFmt

pub trait PrettyFmt: Debug {
    // Required method
    fn pretty_fmt(&self, writer: &mut impl Write) -> Result;
}

Trait for types that can be formatted in a human-readable way in the anvil-zksync output log.

Required Methods

fn pretty_fmt(&self, writer: &mut impl Write) -> Result

Formats the value in a pretty, colorful, and human-readable way, writing to the provided writer.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.