anvil_zksync_core::formatter

Trait ToPrettyString

pub trait ToPrettyString {
    // Required method
    fn to_string_pretty(&self) -> String;
}

Extension trait that provides a convenient method to get a pretty-formatted string.

This trait is automatically implemented for all types that implement PrettyFmt.

Required Methods

fn to_string_pretty(&self) -> String

Converts the value to a pretty-formatted string.