anvil_zksync_core

Module node

anvil-zksync, that supports forking other networks.

Modules

blockchain

diagnostics

Routines to help with node diagnostics, parsing, extracting information from different subsystems. The [crate::formatter] module is used to output it.

error

fork

node_executor

time

Structs

BlockSealer

ImpersonationManager

Manages impersonated accounts across the system.

InMemoryNode

In-memory node, that can be used for local & unit testing. It also supports the option of forking testnet/mainnet. All contents are removed when object is destroyed.

InMemoryNodeInner

Helper struct for InMemoryNode.

NodeExecutor

Snapshot

Creates a restorable snapshot for the InMemoryNodeInner. The snapshot contains all the necessary data required to restore the InMemoryNodeInner state to a previous point in time.

TestNodeFeeInputProvider

TransactionResult

TxBatch

A batch of transactions meant to be sealed as a block. All transactions in the batch share the same impersonation status on the moment of the batch’s creation.

TxExecutionInfo

Information about the executed transaction.

TxPool

ZkSyncOSHelpers

Enums

BlockSealerMode

Represents different modes of block sealing available on the node

StorageKeyLayout

VersionedState

Constants

ESTIMATE_GAS_ACCEPTABLE_OVERESTIMATION

Acceptable gas overestimation limit.

MAX_PREVIOUS_STATES

The maximum number of previous blocks to store the state for.

MAX_TX_SIZE

Max possible size of an ABI encoded tx (in bytes). NOTE: this deviates slightly from the default value in the main node config, that being api.max_tx_size = 1_000_000.

PROTOCOL_VERSION

The zks protocol version.

Functions

compute_hash

create_block

create_genesis

create_genesis_from_json

load_last_l1_batch