anvil_zksync_core::node::blockchain

Trait ReadBlockchain

pub trait ReadBlockchain:
    Send
    + Sync
    + Debug {

    // Required methods
    fn dyn_cloned(&self) -> Box<dyn ReadBlockchain>;
    fn protocol_version(&self) -> ProtocolVersionId;
    fn current_batch<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = L1BatchNumber> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn current_block_number<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = L2BlockNumber> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn current_block_hash<'life0, 'async_trait>(
        &'life0 self,
    ) -> Pin<Box<dyn Future<Output = H256> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_block_by_hash<'life0, 'life1, 'async_trait>(
        &'life0 self,
        hash: &'life1 H256,
    ) -> Pin<Box<dyn Future<Output = Option<Block<TransactionVariant>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_block_by_number<'life0, 'async_trait>(
        &'life0 self,
        number: L2BlockNumber,
    ) -> Pin<Box<dyn Future<Output = Option<Block<TransactionVariant>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_block_by_id<'life0, 'async_trait>(
        &'life0 self,
        block_id: BlockId,
    ) -> Pin<Box<dyn Future<Output = Option<Block<TransactionVariant>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_block_hash_by_number<'life0, 'async_trait>(
        &'life0 self,
        number: L2BlockNumber,
    ) -> Pin<Box<dyn Future<Output = Option<H256>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_block_hash_by_id<'life0, 'async_trait>(
        &'life0 self,
        block_id: BlockId,
    ) -> Pin<Box<dyn Future<Output = Option<H256>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_block_number_by_hash<'life0, 'life1, 'async_trait>(
        &'life0 self,
        hash: &'life1 H256,
    ) -> Pin<Box<dyn Future<Output = Option<L2BlockNumber>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_block_number_by_id<'life0, 'async_trait>(
        &'life0 self,
        block_id: BlockId,
    ) -> Pin<Box<dyn Future<Output = Option<L2BlockNumber>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_block_tx_hashes_by_number<'life0, 'async_trait>(
        &'life0 self,
        number: L2BlockNumber,
    ) -> Pin<Box<dyn Future<Output = Option<Vec<H256>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_block_tx_hashes_by_id<'life0, 'async_trait>(
        &'life0 self,
        block_id: BlockId,
    ) -> Pin<Box<dyn Future<Output = Option<Vec<H256>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_block_tx_by_id<'life0, 'async_trait>(
        &'life0 self,
        block_id: BlockId,
        index: usize,
    ) -> Pin<Box<dyn Future<Output = Option<Transaction>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_block_tx_count_by_id<'life0, 'async_trait>(
        &'life0 self,
        block_id: BlockId,
    ) -> Pin<Box<dyn Future<Output = Option<usize>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_block_details_by_number<'life0, 'async_trait>(
        &'life0 self,
        number: L2BlockNumber,
        l2_fair_gas_price: u64,
        fair_pubdata_price: Option<u64>,
        base_system_contracts_hashes: BaseSystemContractsHashes,
    ) -> Pin<Box<dyn Future<Output = Option<BlockDetails>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_tx_receipt<'life0, 'life1, 'async_trait>(
        &'life0 self,
        tx_hash: &'life1 H256,
    ) -> Pin<Box<dyn Future<Output = Option<TransactionReceipt>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_tx_debug_info<'life0, 'life1, 'async_trait>(
        &'life0 self,
        tx_hash: &'life1 H256,
        only_top: bool,
    ) -> Pin<Box<dyn Future<Output = Option<DebugCall>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_tx_api<'life0, 'life1, 'async_trait>(
        &'life0 self,
        tx_hash: &'life1 H256,
    ) -> Pin<Box<dyn Future<Output = Result<Option<Transaction>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_detailed_tx<'life0, 'async_trait>(
        &'life0 self,
        tx: Transaction,
    ) -> Pin<Box<dyn Future<Output = Option<DetailedTransaction>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_tx_details<'life0, 'life1, 'async_trait>(
        &'life0 self,
        tx_hash: &'life1 H256,
    ) -> Pin<Box<dyn Future<Output = Option<TransactionDetails>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_zksync_tx<'life0, 'life1, 'async_trait>(
        &'life0 self,
        tx_hash: &'life1 H256,
    ) -> Pin<Box<dyn Future<Output = Option<Transaction>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_filter_logs<'life0, 'life1, 'async_trait>(
        &'life0 self,
        log_filter: &'life1 LogFilter,
    ) -> Pin<Box<dyn Future<Output = Vec<Log>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
    fn get_batch_header<'life0, 'async_trait>(
        &'life0 self,
        batch_number: L1BatchNumber,
    ) -> Pin<Box<dyn Future<Output = Option<L1BatchHeader>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_batch_state_diffs<'life0, 'async_trait>(
        &'life0 self,
        batch_number: L1BatchNumber,
    ) -> Pin<Box<dyn Future<Output = Option<Vec<StateDiffRecord>>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_batch_aggregation_root<'life0, 'async_trait>(
        &'life0 self,
        batch_number: L1BatchNumber,
    ) -> Pin<Box<dyn Future<Output = Option<H256>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_raw_transaction<'life0, 'async_trait>(
        &'life0 self,
        tx_hash: H256,
    ) -> Pin<Box<dyn Future<Output = Option<Bytes>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;
    fn get_raw_transactions<'life0, 'async_trait>(
        &'life0 self,
        block_number: BlockId,
    ) -> Pin<Box<dyn Future<Output = Vec<Bytes>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait;

}

Read-only view on blockchain state.

Required Methods

fn dyn_cloned(&self) -> Box<dyn ReadBlockchain>

Alternative for Clone::clone that is object safe.

fn protocol_version(&self) -> ProtocolVersionId

Current protocol version used by the chain.

fn current_batch<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = L1BatchNumber> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns last sealed batch’s number. At least one sealed batch is guaranteed to be present in the storage at any given time.

fn current_block_number<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = L2BlockNumber> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns last sealed block’s number. At least one sealed block is guaranteed to be present in the storage at any given time.

fn current_block_hash<'life0, 'async_trait>( &'life0 self, ) -> Pin<Box<dyn Future<Output = H256> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Returns last sealed block’s hash. At least one sealed block is guaranteed to be present in the storage at any given time.

fn get_block_by_hash<'life0, 'life1, 'async_trait>( &'life0 self, hash: &'life1 H256, ) -> Pin<Box<dyn Future<Output = Option<Block<TransactionVariant>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Retrieve full block by its hash. Returns None if no block was found. Note that the block might still be a part of the chain but is available in the fork instead.

fn get_block_by_number<'life0, 'async_trait>( &'life0 self, number: L2BlockNumber, ) -> Pin<Box<dyn Future<Output = Option<Block<TransactionVariant>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieve full block by its number. Returns None if no block was found. Note that the block might still be a part of the chain but is available in the fork instead.

fn get_block_by_id<'life0, 'async_trait>( &'life0 self, block_id: BlockId, ) -> Pin<Box<dyn Future<Output = Option<Block<TransactionVariant>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieve full block by id. Returns None if no block was found. Note that the block might still be a part of the chain but is available in the fork instead.

fn get_block_hash_by_number<'life0, 'async_trait>( &'life0 self, number: L2BlockNumber, ) -> Pin<Box<dyn Future<Output = Option<H256>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieve block hash by its number. Returns None if no block was found. Note that the block might still be a part of the chain but is available in the fork instead.

fn get_block_hash_by_id<'life0, 'async_trait>( &'life0 self, block_id: BlockId, ) -> Pin<Box<dyn Future<Output = Option<H256>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieve block hash by id. Returns None if no block was found. Note that the block might still be a part of the chain but is available in the fork instead.

fn get_block_number_by_hash<'life0, 'life1, 'async_trait>( &'life0 self, hash: &'life1 H256, ) -> Pin<Box<dyn Future<Output = Option<L2BlockNumber>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Retrieve block number by its hash. Returns None if no block was found. Note that the block might still be a part of the chain but is available in the fork instead.

fn get_block_number_by_id<'life0, 'async_trait>( &'life0 self, block_id: BlockId, ) -> Pin<Box<dyn Future<Output = Option<L2BlockNumber>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieve block number by id. Returns None if no block was found. Note that the block might still be a part of the chain but is available in the fork instead.

fn get_block_tx_hashes_by_number<'life0, 'async_trait>( &'life0 self, number: L2BlockNumber, ) -> Pin<Box<dyn Future<Output = Option<Vec<H256>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieve all transactions hashes from a block by its number. Returns None if no block was found. Note that the block might still be a part of the chain but is available in the fork instead.

fn get_block_tx_hashes_by_id<'life0, 'async_trait>( &'life0 self, block_id: BlockId, ) -> Pin<Box<dyn Future<Output = Option<Vec<H256>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieve all transactions hashes from a block by id. Returns None if no block was found. Note that the block might still be a part of the chain but is available in the fork instead.

fn get_block_tx_by_id<'life0, 'async_trait>( &'life0 self, block_id: BlockId, index: usize, ) -> Pin<Box<dyn Future<Output = Option<Transaction>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieve a transaction from a block by id and index of the transaction. Returns None if either no block was found or no transaction exists in the block under that index. Note that the block might still be a part of the chain but is available in the fork instead.

fn get_block_tx_count_by_id<'life0, 'async_trait>( &'life0 self, block_id: BlockId, ) -> Pin<Box<dyn Future<Output = Option<usize>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieve number of transactions in a block by id. Returns None if no block was found. Note that the block might still be a part of the chain but is available in the fork instead.

fn get_block_details_by_number<'life0, 'async_trait>( &'life0 self, number: L2BlockNumber, l2_fair_gas_price: u64, fair_pubdata_price: Option<u64>, base_system_contracts_hashes: BaseSystemContractsHashes, ) -> Pin<Box<dyn Future<Output = Option<BlockDetails>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieve block details (as defined in zks_getBlockDetails) by id. Returns None if no block was found. Note that the block might still be a part of the chain but is available in the fork instead.

fn get_tx_receipt<'life0, 'life1, 'async_trait>( &'life0 self, tx_hash: &'life1 H256, ) -> Pin<Box<dyn Future<Output = Option<TransactionReceipt>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Retrieve transaction receipt by transaction’s hash. Returns None if no transaction was found. Note that the transaction might still be a part of the chain but is available in the fork instead.

fn get_tx_debug_info<'life0, 'life1, 'async_trait>( &'life0 self, tx_hash: &'life1 H256, only_top: bool, ) -> Pin<Box<dyn Future<Output = Option<DebugCall>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Retrieve transaction debug information by transaction’s hash. Returns None if no transaction was found. Note that the transaction might still be a part of the chain but is available in the fork instead.

fn get_tx_api<'life0, 'life1, 'async_trait>( &'life0 self, tx_hash: &'life1 H256, ) -> Pin<Box<dyn Future<Output = Result<Option<Transaction>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Retrieve transaction in API format by transaction’s hash. Returns None if no transaction was found. Note that the transaction might still be a part of the chain but is available in the fork instead.

fn get_detailed_tx<'life0, 'async_trait>( &'life0 self, tx: Transaction, ) -> Pin<Box<dyn Future<Output = Option<DetailedTransaction>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieve detailed transaction (as defined in anvil_mine_detailed) by API transaction. Returns None if no transaction was found. Note that the transaction might still be a part of the chain but is available in the fork instead.

fn get_tx_details<'life0, 'life1, 'async_trait>( &'life0 self, tx_hash: &'life1 H256, ) -> Pin<Box<dyn Future<Output = Option<TransactionDetails>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Retrieve detailed transaction (as defined in zks_getTransactionDetails) by transaction’s hash. Returns None if no transaction was found. Note that the transaction might still be a part of the chain but is available in the fork instead.

fn get_zksync_tx<'life0, 'life1, 'async_trait>( &'life0 self, tx_hash: &'life1 H256, ) -> Pin<Box<dyn Future<Output = Option<Transaction>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Retrieve ZKsync transaction (as defined in zks_getRawBlockTransactions) by transaction’s hash. Returns None if no transaction was found. Note that the transaction might still be a part of the chain but is available in the fork instead.

fn get_filter_logs<'life0, 'life1, 'async_trait>( &'life0 self, log_filter: &'life1 LogFilter, ) -> Pin<Box<dyn Future<Output = Vec<Log>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Retrieve all logs matching given filter. Does not return matching logs from pre-fork blocks.

fn get_batch_header<'life0, 'async_trait>( &'life0 self, batch_number: L1BatchNumber, ) -> Pin<Box<dyn Future<Output = Option<L1BatchHeader>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieve batch header by its number.

fn get_batch_state_diffs<'life0, 'async_trait>( &'life0 self, batch_number: L1BatchNumber, ) -> Pin<Box<dyn Future<Output = Option<Vec<StateDiffRecord>>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieve batch state diffs by its number.

fn get_batch_aggregation_root<'life0, 'async_trait>( &'life0 self, batch_number: L1BatchNumber, ) -> Pin<Box<dyn Future<Output = Option<H256>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieve batch aggregation root by its number.

fn get_raw_transaction<'life0, 'async_trait>( &'life0 self, tx_hash: H256, ) -> Pin<Box<dyn Future<Output = Option<Bytes>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieves raw transaction by its hash.

fn get_raw_transactions<'life0, 'async_trait>( &'life0 self, block_number: BlockId, ) -> Pin<Box<dyn Future<Output = Vec<Bytes>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait,

Retrieves raw transactions from a block by its id or number.